Assignment 4 Playing with Graphical Objects
The program that displays the graphical objects will receive as input a file containing a list of names of
image files, each corresponding to a graphical object. The objects will be rendered on a window and the user
will move some them around using the keyboard.

Including 4 kinds of Objects:
 fixed objects, which cannot move
 objects that can be moved by the user
 objects that are moved by the computer
 target objects, that disappear when the user-controlled objects run into them.

These graphical objects will be part of a “pac-man”-like game. Objects moved by the computer will chase
the user controlled ones. These latter in turn will try to get rid of the target objects. Fixed objects constrain
the movement of the mobile ones