Assignment 2 Blocked Tic-Tac-Toe
Implementation of a Java program that plays blocked (n; k)-tic-tac-toe. Blocked (n; k)-tic-tac-toe is played on a board of size n * n where some positions are blocked or unavailable to the players, and to win the game a player needs to put k symbols in-line, i.e. on adjacent positions of the same row, column, or diagonal. The program will play against a human opponent.